Add new build command for gh-pages specifically, with path fixes
If using a relative path (e.g. "./") for Snowpack baseUrl (which is
then passed to esbuild as publicPath) with splitting enabled, dynamic
imports within the final bundle have incorrect paths.

For example, if importing the module html-to-image, the generated
path is "./_snowpack/pkg/html-to-image.js". In ESM, this is resolved
relative to the module where the import is called, and so in our case
it would try load the file from "<root>/dist/_snowpack/...", when the
correct path would be "<root>/_snowpack/...".

The new gh-pages specific build command replaces the relative baseUrl
with an absolute path, specifically for the final GitHub pages URL,
which avoids this issue.